# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: KAOS API
  description: API de la Plataforma de Gestión de Equipos de Desarrollo
  version: 1.0.0
  contact:
    name: Equipo KAOS
    email: kaos@ehcos.com
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
paths:
...
  /api/v1/squads:
    get:
      operationId: listarSquads
      summary: Lista squads con filtro opcional por estado
      tags: [Squad]
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVO, INACTIVO]
          description: Filtrar por estado del squad
      responses:
        "200":
          description: Lista de squads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SquadResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: crearSquad
      summary: Crea un nuevo squad
      tags: [Squad]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SquadRequest"
            example:
              nombre: "red"
              descripcion: "Squad Red — Correctivos y evolutivos"
              idSquadCorrJira: "CORR-RED"
              idSquadEvolJira: "EVOL-RED"
      responses:
        "201":
          description: Squad creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/squads/{id}:
    get:
      operationId: obtenerSquad
      summary: Obtiene un squad por ID
      tags: [Squad]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Squad encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquadResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: actualizarSquad
      summary: Actualiza un squad
      tags: [Squad]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SquadRequest"
      responses:
        "200":
          description: Squad actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/squads/{id}/desactivar:
    patch:
      operationId: desactivarSquad
      summary: Desactiva un squad (soft delete)
      description: Cambia el estado del squad a INACTIVO. No se permite si tiene sprints activos.
      tags: [Squad]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Squad desactivado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════
  # PERSONAS
  # ═══════════════════════════════════════════════
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Identificador del recurso
...
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SquadRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          description: Nombre del squad
          maxLength: 100
          example: "red"
        descripcion:
          type: string
          description: Descripción del squad
          example: "Squad Red"
        idSquadCorrJira:
          type: string
          description: ID del board de correctivos en Jira
          maxLength: 100
          example: "CORR-RED"
        idSquadEvolJira:
          type: string
          description: ID del board de evolutivos en Jira
          maxLength: 100
          example: "EVOL-RED"
    SquadResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        descripcion:
          type: string
        estado:
          type: string
          enum: [ACTIVO, INACTIVO]
        idSquadCorrJira:
          type: string
        idSquadEvolJira:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- PERSONA ---
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Recurso no encontrado"
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time